# apps/conference/backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: WSO2Con Conference Backend
  version: "1.0.0"
  description: >
    Conference data, attendee search, networking, recommendation and
    AI-assistant endpoints.
    All routes except /health require the `x-jwt-assertion` header.
servers:
  - url: /
security:
  - jwtAssertion: []
...
paths:
...
  /sessions/current:
    get:
      tags: [sessions]
      summary: All sessions in the current conference, ordered by start time
      description: >
        Returns every session in the latest-start_date conference as Session
        objects (the same shape the agenda endpoints use), ordered by start
        time. Includes speaker-less sessions (breaks/registration). Carries
        ETag/Cache-Control validators.
      responses:
        "200":
          description: "OK"
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Session" }
        "304": { description: "Not Modified" }
  /sessions/{id}:
    get:
      tags: [sessions]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "304": { description: "Not Modified" }
        "404": { description: "No such session" }
  /events:
    get:
      tags: [events]
      parameters:
        - name: previous
          in: query
          required: false
          schema: { type: boolean }
          description: "When `true`, returns only non-current events (history modal)."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }
        "304": { description: "Not Modified" }
  /events/current:
    get:
      tags: [events]
      description: >
        The current conference (the one with the latest start_date), as a single
        object rather than an array. Clients read `id` from this and pin later
        requests to it.
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }
        "304": { description: "Not Modified" }
        "404": { description: "No conference is configured" }
  /events/{eventId}/agendas:
    get:
      tags: [events]
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
            pattern: '^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|current)$'
          description: "A conference_config id, or the literal `current`."
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EventAgenda" }
        "304": { description: "Not Modified" }
        "400":
          description: >-
            eventId is neither a UUID nor the literal "current". conference_config.id
            is a uuid column, so this used to reach the query and surface as a 500.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: eventId must be a valid UUID
  /event-agendas:
    get:
      tags: [events]
      description: "Query-param variant of the agenda resource."
      parameters:
        - { name: isCurrent, in: query, required: false, schema: { type: boolean } }
        - name: eventId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        "200":
          description: "OK"
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/EventAgenda" }
        "304": { description: "Not Modified" }
        "400":
          description: >-
            Neither eventId nor isCurrent=true supplied, or the supplied eventId
            is not a UUID (the latter was previously a 500).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: eventId must be a valid UUID
  /speakers:
    get:
      tags: [speakers]
      parameters:
        - name: eventId
          in: query
          required: false
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
          description: "Restrict to speakers with a session in this conference (and show only those sessions)."
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: >-
            Case-insensitive substring search over the speaker's name, title and
            company (all matched server-side after decryption). Company is searched
            but not returned by this endpoint.
      responses:
        "200":
          description: "OK, ordered by name"
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SpeakerSummary" }
        "304": { description: "Not Modified" }
        "400":
          description: >-
            eventId was supplied and is not a UUID. It is optional; the check fires
            only when present.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: eventId must be a valid UUID
  /speakers/{id}:
    get:
      tags: [speakers]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: "OK"
          headers:
            ETag: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Speaker" }
        "304": { description: "Not Modified" }
        "400":
          description: "id is not a UUID (previously a 500 from the cast)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: id must be a valid UUID
        "404": { description: "No such visible speaker" }
  /attendees:
    post:
      tags: [attendees]
      summary: Create or replace the caller's own attendee profile
      description: >
        Both identity fields come from the caller's JWT -- idp_uuid from `sub`,
        email from the `email` claim. A body `email` is accepted by the decoder
        and ignored, so a caller cannot create or overwrite another attendee's
        row. The write is an upsert scoped to the caller's email; when that email
        already carries a row bound to a different idp_uuid the upsert is a no-op
        and the route answers 409 rather than claiming a write that did not happen.
        Returns no body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendeeInsert" }
      responses:
        "201": { description: "Created or updated. No body." }
        "400":
          description: >-
            Malformed JSON body, or a missing/empty firstName or lastName. An empty
            POST used to create an email='' row against the caller's UNIQUE idp_uuid
            and wedge the account permanently.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: invalid request body
        "401":
          description: >-
            No authentication, or the token carries no `email` claim (email is the
            key this row is written under).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: authenticated user has no email claim
        "409":
          description: >-
            The caller's email already has an attendee row bound to a different
            idp_uuid. Nothing was written.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: attendee already registered to another account
        "500":
          description: "Internal error"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    patch:
      tags: [attendees]
      summary: Update the authenticated attendee's own profile
      description: >
        There is no target parameter: the row is selected by the JWT `email`
        claim. The route previously took a required `?email=` query parameter that
        chose both the row to patch and the full decrypted row to return, with no
        comparison against the token; any `?email=` a client still sends is now
        ignored. Nil fields are left unchanged; the patched row is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendeePatch" }
      responses:
        "200":
          description: "The patched row, decrypted"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendee" }
        "400":
          description: "Malformed JSON body"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: "No authentication"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: >-
            The caller has no attendee row. (Before the parameter was dropped this
            meant "no row for the queried email".)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                message: attendee not found
        "500":
          description: "Internal error"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /attendees/me:
    get:
      tags: [attendees]
      summary: The caller's own attendee row
      description: >
        Looked up by the JWT `email` claim. Unlike POST /attendees/search this
        returns the caller's own `memberId` and `qrUri` -- qrUri is the check-in
        credential and belongs only to its owner.
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attendee" }
        "401":
          description: "No authentication"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: "The caller has no attendee row"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /user-profile:
    get:
      tags: [attendees]
      summary: Display profile for the caller (name, avatar, QR)
      description: >
        A projection over the caller's attendee row plus JWT name claims. There is
        no fallback to a JWT-only response: a caller with no attendee row gets 404.
        `imageUrl` is a hardcoded default avatar, never derived from the token.
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401":
          description: "No authentication"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: "The caller has no attendee row"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /attendees/search:
    post:
      tags: [attendees]
      summary: Cursor-paginated attendee search (server-side filtering)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid: { type: string, description: "Optional single-attendee lookup by idp uuid." }
                query: { type: string, description: "Case-insensitive substring over name/company/title (decrypted server-side)." }
                cursor: { type: string, description: "Opaque nextCursor from a prior page." }
                limit:
                  type: integer
                  default: 20
                  maximum: 100
                  description: >-
                    The maximum is clamped, not rejected: the repository caps a
                    larger value at 100 and serves the request, and 0 (or an
                    omitted limit) falls through to the default. Only a negative
                    limit is refused, with a 400.
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendeeSearchResult" }
        "400": { description: "Absent or malformed JSON body, invalid cursor, or negative limit" }
        "401":
          description: "No authentication"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: "Internal error"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  /users/me/connections/{connectionId}:
    delete:
      tags: [connections]
      summary: Decline, withdraw, or remove a connection
      description: >
        One route for all three, callable by **either party** at either state:
        the addressee declining a request, the requester withdrawing one, and
        either side removing an established connection are the same operation.


        The row is deleted, not marked declined. The pair therefore returns to
        having no relationship and can connect again later -- a stored
        `declined` state would never appear in any response yet would silently
        no-op every future request between those two people, because the
        re-request conflicts on the pair's unique index.
      parameters:
        - name: connectionId
          in: path
          required: true
          schema: { type: string, format: uuid }
          description: "`connectionId` from GET /users/me/connections."
      responses:
        "204":
          description: "Deleted. No body."
        "400":
          description: "`connectionId` is not a UUID."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: "No authentication"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: >-
            No such connection, or the caller is not a party to it -- including
            a repeated delete of one already removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "500":
          description: "Internal error"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
...
components:
  securitySchemes:
    jwtAssertion:
      type: apiKey
      in: header
      name: x-jwt-assertion
  schemas:
...
    Event:
      type: object
      required: [id, name, isCurrent, timezone, startDate, endDate]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        isCurrent: { type: boolean }
        timezone: { type: string, description: "IANA venue timezone (conference_config.timezone) the instants are expressed in." }
        startDate:
          type: string
          format: date
          description: >-
            First venue-local day of the conference (conference_config.start_date),
            the same column the isCurrent rule sorts on.
        endDate:
          type: string
          format: date
          description: >-
            Last venue-local day of the conference, inclusive. There is no
            end_date column upstream, so this is the last conference_days.date,
            falling back to startDate for a conference whose days are not entered
            yet. A one-day conference repeats startDate here.
        venueName: { type: string, description: "From conference_config.venue_name; omitted when unset." }
        venueAddress: { type: string, description: "From conference_config.venue_address; omitted when unset." }
    EventAgenda:
      type: object
      required: [id, eventId, timezone, date, sessions]
      properties:
        id: { type: string, format: uuid }
        eventId: { type: string, format: uuid }
        timezone: { type: string }
        name: { type: string }
        date: { type: string, format: date }
        sessions:
          type: array
          items: { $ref: "#/components/schemas/Session" }
...